backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
requests>=2.32.0
gTTS>=2.5.0
pydantic>=2.10.0
python-dotenv>=1.0.0
numpy>=2.2.0